[Refactor] click, submit 이벤트 분리
