+++ src/main.js
@@ -1,34 +1,8 @@
-import { navigate, render } from "./router";
-import {
-  loginAction,
-  logoutAction,
-  profileUpdateAction,
-} from "./actions/userActions";
+import { render } from "./router";
+import { addListeners } from "./listeners";
 
+addListeners();
 window.addEventListener("popstate", render);
-document.addEventListener("DOMContentLoaded", render);
-
-document.addEventListener("click", (e) => {
-  if (e.target.matches("a")) {
-    e.preventDefault();
-    const path = e.target.getAttribute("href");
-    navigate(path);
-  }
-
-  if (e.target.matches("#logout")) {
-    e.preventDefault();
-    logoutAction();
-  }
-});
-
-document.addEventListener("submit", (e) => {
-  if (e.target.matches("#login-form")) {
-    e.preventDefault();
-    loginAction(e);
-  }
-
-  if (e.target.matches("#profile-form")) {
-    e.preventDefault();
-    profileUpdateAction();
-  }
+document.addEventListener("DOMContentLoaded", function () {
+  render();
 });
